adapted docs example to new unified callback, added custom callback example
